feat: support decimal quantities on offer items

+++ app/Data/Listing/ListingOfferItemData.php
@@ -12,7 +12,7 @@ class ListingOfferItemData extends Data
     public function __construct(
         public ?int $id,
         public ?int $listingOfferId,
-        public int $quantity,
+        public float $quantity,
         public ?int $item_id,
         public ItemFormData $item,
     ) {}
